React based Application of my MiniProject1
external tools used are MUI library and chartjs.
An example of useEffect is in PatientSummary.jsx that retrieves formdata from local storage
React Hooks used are state, effect, navigate & theme.

Planning to extend the features of the forms that can be used for multiple members of the family and not just the user (for example: save records for spouse/kids/pets)
also a way to communicate with health care providers.